new version without queue and size_t

+++ src/main2.cpp
@@ -11,10 +11,10 @@ T gen() {
 
 int main() {
 	srand(time(NULL));
-	const size_t n = 10;
+	const int n = 10;
 	int* arr;
 	arr = createArr<int, n>(gen);
-	for (size_t i = 0; i < n; i++) {
+	for (int i = 0; i < n; i++) {
 		cout << arr[i] << " ";  
 	}
 	cout << endl;
